Refactors /functions to work with the new query names #127
