+++ react-backend/functions/checkExpire.js
@@ -1,15 +1,13 @@
-const checkExpire = require('../database/queries/CBqueries/checkExpire');
+const checkExpire = require('../database/queries/cb/pwd_token_expire');
 
-module.exports = token => {
-  return new Promise((resolve, reject) => {
+module.exports = token =>
+  new Promise((resolve, reject) => {
     checkExpire(token, (error, result) => {
       if (error) {
         console.log('error from checkExpire ', error);
         reject(error);
       } else {
-        console.log('checkExpire: ', result.rows[0].exists);
         resolve(result.rows[0].exists);
       }
     });
   });
-};
